Add use sections hook for journey
